Add SpacePirateController
Today is the beginning of creating the enemy bot logic. I really like the design I have come up with; I've been playing with this idea for the past few days trying to find something that gonna give the behavior I'm looking for. Created a prefab for the bots and made good progress on the pathfinding algorithm. Only issue is apparently Unity's C#/.NET doesn't provide support for priority queues, so tomorrow's plan is just cope and do extra work lmao